updated word document template
Updated the word template for unique search fields so nothing unintentional in the doc  gets replaced (e.g another name field gets overwritten accidentally), a copy of the template is created in the createdWordDocs folder then changes applied before the file is renamed the company name.

Currently fields are just updated with default values for now.

You need to have UiPath.Word.Activities package installed and this repository should be cloned to C: Drive

If theres an error delete the word document created in the createdworddocs folder and retry.